Move migrate lifecycle code into migrate_lifecycle.go
